model-service/requirements.txt
grpcio 
grpcio-tools
pymongo
python-dotenv
sentence-transformers
openai
torch
transformers